requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
pydantic
mongoengine
python-dotenv
python-multipart
